+++ modules/config/options.lua
@@ -15,48 +15,79 @@ along with the addon. If not, see <http://www.gnu.org/licenses/gpl-3.0.txt>.
 This file is part of Scrap.
 --]]
 
-local Patrons = {{title='Jenkins',people={'Gnare','Arriana Sylvester','SirZooro','ProfessahX'}},{},{title='Ambassador',people={'Sembiance','Fernando Bandeira','Michael Irving','Julia F','Peggy Webb','Lolari','Craig Falb','Mary Barrentine','Grey Sample','Patryk Kalis','Lifeprayer','Steve Lund'}}} -- generated patron list
+local Sushi = LibStub('Sushi-3.1')
+local Options = Scrap:NewModule('Options', Sushi.OptionsGroup('Scrap |TInterface/Addons/Scrap/art/enabled-icon:13:13:0:0|t'))
 local L = LibStub('AceLocale-3.0'):GetLocale('Scrap')
 
-local Options = SushiMagicGroup(Scrap.options)
-Options:SetAddon('Scrap')
-Options:SetFooter('Copyright 2008-2019 João Cardoso')
-Options:SetTitle(Scrap.options.name)
-Options:SetChildren(function(self)
-	self:CreateHeader('Behaviour', 'GameFontHighlight', true)
-	self:Create('CheckButton', 'sell')
-	self:Create('CheckButton', 'repair')
-	self:Create('CheckButton', 'guild', nil, not Scrap.sets.repair, true)
-	self:Create('CheckButton', 'safe')
-	self:Create('CheckButton', 'learn')
+local PATRONS = {{title='Jenkins',people={'Gnare','Arriana Sylvester','SirZooro','ProfessahX'}},{},{title='Ambassador',people={'Sembiance','Fernando Bandeira','Michael Irving','Julia F','Peggy Webb','Lolari','Craig Falb','Mary Barrentine','Grey Sample','Patryk Kalis','Lifeprayer','Steve Lund'}}} -- generated patron list
+local FOOTER = 'Copyright 2008-2019 João Cardoso'
 
-	self:CreateHeader('Filters', 'GameFontHighlight', true)
-	self:Create('CheckButton', 'unusable')
-	self:Create('CheckButton', 'equip')
-	self:Create('CheckButton', 'consumable')
 
-	self:CreateHeader('Visuals', 'GameFontHighlight', true)
-	self:Create('CheckButton', 'glow')
-	self:Create('CheckButton', 'icons')
-end)
+--[[ Startup ]]--
 
-local Share = SushiCheckButton(Options)
-Share:SetPoint('BOTTOM', Options, 'TOP', 115, 10)
-Share:SetText(L.CharSpecific)
-Share:SetChecked(not Scrap.charsets.shared)
-Share:SetScale(.9)
-Share:SetCall('OnInput', function(self, v)
-	Scrap.charsets.share = not v
-	Scrap:SendSignal('SETS_CHANGED')
-end)
+function Options:OnEnable()
+	local credits = Sushi.CreditsGroup(self, PATRONS)
+	credits:SetSubtitle(nil, 'http://www.patreon.com/jaliborc')
+	credits:SetFooter(FOOTER)
 
-local Credits = SushiCreditsGroup:CreateOptionsCategory(Options:GetTitle())
-Credits:SetFooter('Copyright 2008-2019 João Cardoso')
-Credits:SetWebsite('http://www.patreon.com/jaliborc')
-Credits:SetPeople(Patrons)
-Credits:SetAddon('Scrap')
+	local share = Sushi.Check(self)
+	share:SetChecked(not Scrap_CharSets.shared)
+	share:SetPoint('TOPRIGHT', -30, 36)
+	share:SetText(L.CharSpecific)
+	share:SetScale(.9)
+	share:SetCall('OnInput', function(share, v)
+		Scrap_CharSets.share = not v
+		self:SendSignal('SETS_CHANGED')
+	end)
 
-function Scrap.options.default()
+	self:SetFooter(FOOTER)
+	self:SetSubtitle(L.Description)
+	self:SetChildren(self.Populate)
+	self:SetCall('OnDefaults', self.OnDefaults)
+end
+
+function Options:OnDefaults()
 	Scrap_Sets, Scrap_CharSets = nil
-	ReloadUI()
+	self:SendSignal('SETS_CHANGED')
+	self:Update()
+end
+
+function Options:Populate()
+	self:AddHeader('Behaviour')
+	self:AddCheck {set = 'sell', text = 'AutoSell'}
+	self:AddCheck {set = 'repair', text = 'AutoRepair'}
+	self:AddCheck {set = 'guild', text = 'GuildRepair', parent = 'repair'}
+	self:AddCheck {set = 'safe', text = 'SafeMode'}
+	self:AddCheck {set = 'destroy', text = 'DestroyWorthless'}
+
+	self:AddHeader(CALENDAR_FILTERS)
+	self:AddCheck {set = 'unusable', text = 'Unusable', char = true}
+	self:AddCheck {set = 'equip', text = 'LowEquip', char = true}
+	self:AddCheck {set = 'consumable', text = 'LowConsume', char = true}
+	self:AddCheck {set = 'learn', text = 'Learning'}
+
+	self:AddHeader('Visuals')
+	self:AddCheck {set = 'glow', text = 'Glow'}
+	self:AddCheck {set = 'icons', text = 'Icons'}
+end
+
+
+--[[ API ]]--
+
+function Options:AddHeader(text)
+	self:Add('Header', text, GameFontHighlight, true)
+end
+
+function Options:AddCheck(info)
+	local sets = info.char and Scrap_CharSets or Scrap_Sets
+	local b = self:Add('Check', L[info.text])
+	b.left = b.left + (info.parent and 10 or 0)
+	b:SetEnabled(not info.parent or sets[info.parent])
+	b:SetTip(L[info.text], L[info.text .. 'Tip'])
+	b:SetChecked(sets[info.set])
+	b:SetSmall(info.parent)
+	b:SetCall('OnInput', function(b, v)
+		sets[info.set] = v
+		self:SendSignal('LIST_CHANGED')
+	end)
 end