+++ modules/merchant/learning.lua
@@ -15,69 +15,81 @@ along with the addon. If not, see <http://www.gnu.org/licenses/gpl-3.0.txt>.
 This file is part of Scrap.
 --]]
 
-hooksecurefunc('UseContainerItem', function(...)
-  self:OnItemSold(...)
-end)
+local Learn = Scrap:NewModule('Learning') -- a really really dumb ml algortihm
 
-local buyBack = BuybackItem
-BuybackItem = function(...)
-  self:OnItemRefund(...)
-  return buyBack(...)
+
+--[[ Events ]] --
+
+function Learn:OnEnable()
+  hooksecurefunc('UseContainerItem', function(...)
+    if self:IsActive() then
+      self:OnItemSold(...)
+    end
+  end)
+
+  local buyBack = BuybackItem
+  BuybackItem = function(...)
+    if self:IsActive() then
+      self:OnItemRefund(...)
+    end
+    return buyBack(...)
+  end
 end
 
+function Learn:OnItemSold(...)
+	local id = GetContainerItemID(...)
+	if not id or Scrap.junk[id] ~= nil or Scrap:CheckFilters(id, ...) then
+		return
+	end
 
---[[ AI ]]--
+	local stack = select(2, GetContainerItemInfo(...))
+	if GetItemCount(id, true) == stack then
+		local link = GetContainerItemLink(...)
+		local maxStack = select(8, GetItemInfo(id))
+		local stack = self:GetBuypackStack(link) + stack
 
-function Button:OnItemSold (...)
-	if Scrap.sets.learn and self:IsVisible() and not self.selling then
-		local id = GetContainerItemID(...)
-		if not id or self.Junk[id] ~= nil or self:CheckFilters(id, ...) then
-			return
-		end
+		local old = Scrap.charsets.ml[id] or 0
+		local new = old + stack / maxStack
+		Scrap.charsets.ml[id] = new
 
-		local stack = select(2, GetContainerItemInfo(...))
-		if GetItemCount(id, true) == stack then
-			local link = GetContainerItemLink(...)
-			local maxStack = select(8, GetItemInfo(id))
-			local stack = self:GetBuypackStack(link) + stack
+    if old < 2 and new >= 2 then
+      Scrap:Print(L.Added, link, 'LOOT')
+      Scrap:SendSignal('LIST_CHANGED', id)
+    end
+	end
+end
 
-			local old = Scrap.charsets.ml[id] or 0
-			local new = old + stack / maxStack
+function Learn:OnItemRefund(index)
+	local link = GetBuybackItemLink(index)
+	local id = ink and tonumber(link:match('item:(%d+)'))
+	local old = Scrap.charsets.ml[id]
 
-			if old < 2 and new >= 2 then
-				self:Print(L.Added, link, 'LOOT')
-			end
+	if old then
+		local maxStack = select(8, GetItemInfo(id))
+		local stack = self:GetBuypackStack(link)
 
+		local new = old - stack / maxStack
+		if new <= 0 then
+			Scrap.charsets.ml[id] = nil
+		else
 			Scrap.charsets.ml[id] = new
 		end
+
+    if old >= 2 and new < 2 then
+      Scrap:Print(L.Removed, link, 'LOOT')
+      Scrap:SendSignal('LIST_CHANGED', id)
+    end
 	end
 end
 
-function Button:OnItemRefund (index)
-	if Scrap.sets.learn then
-		local link = GetBuybackItemLink(index)
-		local id = ink and tonumber(link:match('item:(%d+)'))
-		local old = Scrap.charsets.ml[id]
 
-		if old then
-			local maxStack = select(8, GetItemInfo(id))
-			local stack = self:GetBuypackStack(link)
+--[[ API ]]--
 
-			local new = old - stack / maxStack
-			if old >= 2 and new < 2 then
-				self:Print(L.Removed, link, 'LOOT')
-			end
-
-			if new <= 0 then
-				Scrap.charsets.ml[id] = nil
-			else
-				Scrap.charsets.ml[id] = new
-			end
-		end
-	end
+function Learn:IsActive()
+  return Scrap.sets.learn and MerchantFrame:IsVisible()
 end
 
-function Button:GetBuypackStack (link)
+function Learn:GetBuypackStack(link)
 	local stack = 0
 	for i = 1, GetNumBuybackItems() do
 		if GetBuybackItemLink(i) == link then