+++ Changelog.md
@@ -2,6 +2,7 @@
 * Changed when low level gray equipment is classified as junk by default.
 * Added new items to default junk exclusion as requested by users.
 * Added options to not destroy items with no sale value at the vendor.
+* Added keybinding to destroy junk items.
 * Complete internal rewrite of settings, variables and file internal structure.
    * `Scrap_Visualizer` is not part of `Scrap_Merchant`
    * Renamed `Scrap_Options` to `Scrap_Config`